packages/cli: move plugin diff modules to separate diff lib
